Moved FPLite extensions to separate library
